+++ projetocampeonato/ClubeCopaBrasil.java
@@ -2,7 +2,7 @@ package projetocampeonato;
 
 public class ClubeCopaBrasil extends Clube {
     
-    public ClubeCopaBrasil(String nome, int fundacao, String local, double torcida, int score) {
+    public ClubeCopaBrasil(String nome, int fundacao, String local, long torcida, int score) {
         super(nome, fundacao, local, torcida, score);
     }
 }
